Fix 'Slowly add and remove a track' test
