+++ e2e/scenarios/raceCondition.spec.ts
@@ -1,5 +1,15 @@
 import { expect, Page, test, TestInfo } from "@playwright/test";
-import { assertThatOtherVideoIsPlaying, createAndJoinPeer, createRoom, takeScreenshot } from "./utils";
+import {
+  addAndRemoveTrack,
+  addAndReplaceTrack,
+  addBothMockTracks,
+  assertThatAllTracksAreReady, assertThatBothTrackAreDifferent,
+  assertThatOtherVideoIsPlaying, assertThatTrackBackgroundColorIsOk, assertThatTrackReplaceStatusIsSuccess, clickButton,
+  createAndJoinPeer,
+  createRoom, removeTrack,
+  takeScreenshot,
+  TO_PASS_CONFIG
+} from "./utils";
 
 test.afterEach(async ({ context }, testInfo) => {
   for (const page of context.pages()) {
@@ -126,7 +136,7 @@ test("RC: Add 2 tracks at the same time and remove one track", async ({ page: se
   });
 });
 
-test("Add and replace tracks (slow)", async ({ page: senderPage, context }) => {
+test("Slowly add and replace tracks", async ({ page: senderPage, context }) => {
   // given
   await senderPage.goto("/");
   const roomId = await createRoom(senderPage);
@@ -152,7 +162,7 @@ test("Add and replace tracks (slow)", async ({ page: senderPage, context }) => {
   await assertThatTrackReplaceStatusIsSuccess(senderPage, "success");
 });
 
-test("RC: Add and replace a track (fast)", async ({ page: senderPage, context }, testInfo) => {
+test("RC: Quickly add and replace a track", async ({ page: senderPage, context }, testInfo) => {
   // given
   await senderPage.goto("/");
   const roomId = await createRoom(senderPage);
@@ -178,7 +188,7 @@ test("RC: Add and replace a track (fast)", async ({ page: senderPage, context },
   await takeScreenshot(receiverPage, testInfo);
 });
 
-test("Add and remove a track (slow)", async ({ page: senderPage, context }, testInfo) => {
+test("Slowly add and remove a track", async ({ page: senderPage, context }, testInfo) => {
   // given
   await senderPage.goto("/");
   const roomId = await createRoom(senderPage);
@@ -188,21 +198,24 @@ test("Add and remove a track (slow)", async ({ page: senderPage, context }, test
   const receiverPage = await context.newPage();
   await receiverPage.goto("/");
   await createAndJoinPeer(receiverPage, roomId);
-  await receiverPage.waitForTimeout(1000);
 
   // when
   await clickButton(senderPage, "Add a heart");
-  await senderPage.waitForTimeout(1000);
-  await clickButton(senderPage, "Remove a heart");
 
   // then
   await assertThatAllTracksAreReady(receiverPage, senderId, 1);
+
+  // when
+  await senderPage.waitForTimeout(1000)
+  await clickButton(senderPage, "Remove a heart");
+
+  // then
   await assertThatAllTracksAreReady(receiverPage, senderId, 0);
 
   await takeScreenshot(receiverPage, testInfo);
 });
 
-test.skip("RC: Add and remove a track (fast)", async ({ page: senderPage, context }, testInfo) => {
+test.skip("RC: Quickly add and remove a track", async ({ page: senderPage, context }, testInfo) => {
   // given
   await senderPage.goto("/");
   const roomId = await createRoom(senderPage);
@@ -223,85 +236,3 @@ test.skip("RC: Add and remove a track (fast)", async ({ page: senderPage, contex
 
   await takeScreenshot(receiverPage, testInfo);
 });
-
-export const addAndReplaceTrack = async (page: Page) =>
-  await test.step("Add and replace track", async () =>
-    await page
-      .getByRole("button", {
-        name: "Add and replace a heart",
-        exact: true,
-      })
-      .click());
-
-export const addAndRemoveTrack = async (page: Page) =>
-  await test.step("Add and remove track", async () =>
-    await page
-      .getByRole("button", {
-        name: "Add and remove a heart",
-        exact: true,
-      })
-      .click());
-
-export const addBothMockTracks = async (page: Page) =>
-  await test.step("Add both tracks", async () =>
-    await page
-      .getByRole("button", {
-        name: "Add both",
-        exact: true,
-      })
-      .click());
-
-export const assertThatAllTracksAreReady = async (page: Page, otherClientId: string, tracks: number) =>
-  await test.step(`Assert that all (${tracks}) tracks are ready`, async () =>
-    await expect(async () =>
-      expect((await page.locator(`div[data-endpoint-id="${otherClientId}"]`).all()).length).toBe(tracks),
-    ).toPass());
-
-export const assertThatTrackBackgroundColorIsOk = async (page: Page, otherClientId: string, color: string) =>
-  await test.step(`Assert that track background color is ${color}`, async () =>
-    await expect(async () =>
-      expect(
-        page.locator(`xpath=//div[@data-endpoint-id="${otherClientId}"]//div[@data-color-name="${color}"]`),
-      ).toBeVisible(),
-    ).toPass());
-
-export const assertThatTrackReplaceStatusIsSuccess = async (page: Page, replaceStatus: string) =>
-  await test.step(`Assert that track background color is ${replaceStatus}`, async () =>
-    await expect(async () =>
-      expect(page.locator(`xpath=//span[@data-replace-status="${replaceStatus}"]`)).toBeVisible(),
-    ).toPass());
-
-export const assertThatTrackIdIsNotEmpty = async (page: Page, locator: string) =>
-  await test.step("Assert that track id is not empty", async () =>
-    await expect(async () => {
-      expect((await page.locator(locator).textContent())?.trim()?.length ?? 0).toBeGreaterThan(0);
-    }).toPass());
-
-export const assertThatBothTrackAreDifferent = async (page: Page, testInfo: TestInfo, name?: string) => {
-  await test.step("Assert that both tracks are different", async () => {
-    const locator1 = `(//div[@data-name="stream-id"])[1]`;
-    const locator2 = `(//div[@data-name="stream-id"])[2]`;
-
-    await assertThatTrackIdIsNotEmpty(page, locator1);
-    await assertThatTrackIdIsNotEmpty(page, locator2);
-
-    await takeScreenshot(page, testInfo, name);
-
-    const text1 = await page.locator(locator1).textContent();
-    const text2 = await page.locator(locator2).textContent();
-
-    expect(text1 !== "").toBe(true);
-    expect(text2 !== "").toBe(true);
-    expect(text1 !== text2).toBe(true);
-  });
-};
-
-export const clickButton = async (page: Page, name: string) =>
-  await test.step(`Add '${name}' track`, async () => {
-    await page.getByRole("button", { name: name, exact: true }).click();
-  });
-
-export const removeTrack = async (page: Page, button: string) =>
-  await test.step(`Remove track ${button}`, async () => {
-    await page.getByRole("button", { name: button, exact: true }).click();
-  });
