+++ e2e/scenarios/utils.ts
@@ -1,6 +1,9 @@
 import { expect, Page, test, TestInfo } from "@playwright/test";
 import { v4 as uuidv4 } from "uuid";
 
+export const TO_PASS_TIMEOUT_MILLIS = 10 * 1000; // 10 seconds
+export const TO_PASS_CONFIG = { timeout: TO_PASS_TIMEOUT_MILLIS };
+
 export const addScreenShare = async (page: Page) =>
   await test.step("Add screenshare", async () => {
     await page.getByRole("button", { name: "Start screenshare", exact: true }).click();
@@ -87,7 +90,7 @@ export const assertThatOtherVideoIsPlaying = async (page: Page) => {
         return 0;
       });
     const firstMeasure = await getDecodedFrames();
-    await expect(async () => expect((await getDecodedFrames()) > firstMeasure).toBe(true)).toPass();
+    await expect(async () => expect((await getDecodedFrames()) > firstMeasure).toBe(true)).toPass(TO_PASS_CONFIG);
   });
 };
 
@@ -119,3 +122,85 @@ export const createPeer = async (page: Page, roomId: string, enableSimulcast: bo
       },
     });
   });
+
+export const clickButton = async (page: Page, name: string) =>
+    await test.step(`Click '${name}' button`, async () => {
+        await page.getByRole("button", { name: name, exact: true }).click();
+    });
+
+export const removeTrack = async (page: Page, button: string) =>
+    await test.step(`Remove track ${button}`, async () => {
+        await page.getByRole("button", { name: button, exact: true }).click();
+    });
+
+export const addAndReplaceTrack = async (page: Page) =>
+    await test.step("Add and replace track", async () =>
+        await page
+            .getByRole("button", {
+                name: "Add and replace a heart",
+                exact: true,
+            })
+            .click());
+
+export const addAndRemoveTrack = async (page: Page) =>
+    await test.step("Add and remove track", async () =>
+        await page
+            .getByRole("button", {
+                name: "Add and remove a heart",
+                exact: true,
+            })
+            .click());
+
+export const addBothMockTracks = async (page: Page) =>
+    await test.step("Add both tracks", async () =>
+        await page
+            .getByRole("button", {
+                name: "Add both",
+                exact: true,
+            })
+            .click());
+
+export const assertThatAllTracksAreReady = async (page: Page, otherClientId: string, tracks: number) =>
+    await test.step(`Assert that all (${tracks}) tracks are ready`, async () =>
+        await expect(async () =>
+            expect((await page.locator(`div[data-endpoint-id="${otherClientId}"]`).all()).length).toBe(tracks),
+        ).toPass(TO_PASS_CONFIG));
+
+export const assertThatTrackBackgroundColorIsOk = async (page: Page, otherClientId: string, color: string) =>
+    await test.step(`Assert that track background color is ${color}`, async () =>
+        await expect(async () =>
+            expect(
+                page.locator(`xpath=//div[@data-endpoint-id="${otherClientId}"]//div[@data-color-name="${color}"]`),
+            ).toBeVisible(),
+        ).toPass(TO_PASS_CONFIG));
+
+export const assertThatTrackReplaceStatusIsSuccess = async (page: Page, replaceStatus: string) =>
+    await test.step(`Assert that track background color is ${replaceStatus}`, async () =>
+        await expect(async () =>
+            expect(page.locator(`xpath=//span[@data-replace-status="${replaceStatus}"]`)).toBeVisible(),
+        ).toPass(TO_PASS_CONFIG));
+
+export const assertThatTrackIdIsNotEmpty = async (page: Page, locator: string) =>
+    await test.step("Assert that track id is not empty", async () =>
+        await expect(async () => {
+            expect((await page.locator(locator).textContent())?.trim()?.length ?? 0).toBeGreaterThan(0);
+        }).toPass(TO_PASS_CONFIG));
+
+export const assertThatBothTrackAreDifferent = async (page: Page, testInfo: TestInfo, name?: string) => {
+    await test.step("Assert that both tracks are different", async () => {
+        const locator1 = `(//div[@data-name="stream-id"])[1]`;
+        const locator2 = `(//div[@data-name="stream-id"])[2]`;
+
+        await assertThatTrackIdIsNotEmpty(page, locator1);
+        await assertThatTrackIdIsNotEmpty(page, locator2);
+
+        await takeScreenshot(page, testInfo, name);
+
+        const text1 = await page.locator(locator1).textContent();
+        const text2 = await page.locator(locator2).textContent();
+
+        expect(text1 !== "").toBe(true);
+        expect(text2 !== "").toBe(true);
+        expect(text1 !== text2).toBe(true);
+    });
+};
